LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/fetch.sub-expected.txt


PASS http->https fetch (cross-scheme => cross-site)
PASS http->http fetch (non-trustworthy destination => no metadata)
PASS http->https fetch (cross-scheme => cross-site): sec-fetch-dest
PASS http->https fetch (cross-scheme => cross-site): sec-fetch-mode
PASS http->https fetch (cross-scheme => cross-site): sec-fetch-site
PASS http->https fetch (cross-scheme => cross-site): sec-fetch-user
FAIL http->http fetch (non-trustworthy destination => no metadata): sec-fetch-dest assert_equals: expected "" but got "empty"
FAIL http->http fetch (non-trustworthy destination => no metadata): sec-fetch-mode assert_equals: expected "" but got "cors"
FAIL http->http fetch (non-trustworthy destination => no metadata): sec-fetch-site assert_equals: expected "" but got "same-origin"
PASS http->http fetch (non-trustworthy destination => no metadata): sec-fetch-user
